Events handling
Made the whole event handling work. Now an event can be created, named and will get a unique ID. Events can be deleted from the listing and the last deletion can be undone as well. Events can be opened for the next step 'participants'.
The only thing missing here is, they aren't saved to the file yet...